bench/requirements.txt
psycopg2-binary
requests
tiktoken
matplotlib
